benchmarks/requirements.txt
linear-crf-torch
TorchCRF
pytorchcrf
pytorch-crf
